Update 26/08/2021: First creation, adding files to repository
I started the project. I had this idea in my head for a long time because it intrigued me, but after a TAFE group I was in voted to create a game with a destructible environment, I decided now was the perfect time to try it.
The first task I did was to make the essential functioning elements for chunk generation:
•	A ‘Block’ struct for storing information about a block in a particular grid coordinate
•	A ‘Chunk’ class for compartmentalising and managing groups of blocks in a 3D grid (with a default of 16x16x16 pixels)
•	A code for generating a mesh based on the blocks in the chunk
•	A ‘BlockData’ ScriptableObject for different block types, such as their textures, durability and other properties.
This all turned out reasonably well, but it was far more difficult than I expected. The biggest challenge was determining efficient ways to check angles and generate the appropriate faces, most notably determining and ordering the appropriate corner positions for the various faces that needed to be rendered. After much brain-wracking, the solution I came up with was to generate four manual Vector3 values representing the corners of the front face, then create a Quaternion using the desired direction and use it to rotate the corners so they face in the appropriate direction.
This worked incredibly well, as the four corners are always in the same position relative to each other, meaning their order and the indexes used to reference them for face generation don’t have to be changed (except for increasing the indexes to match the amount of corner values in the mesh so they reference the new corners rather than the first set in the array). It will also be very easy to apply additional rotations if I add blocks that face a specific direction, such as doors or stairs. After dropping in some physics objects, it was cathartic to see that the mesh and collider had generated appropriately and physics objects were realistically reacting to them.
So that’s nice. Unfortunately, there is a lot more work to do. I need to implement the following important features:
•	Adding different textures to the chunk mesh based on each different block
•	A level manager that generates several chunks and manages changes between them so it appears seamless
•	The ability to, when affecting mesh inside a chunk, to easily check the data of the chunk next to it
•	Obtaining block info from a world position without specifying a chunk
•	Saving chunk data to a string
•	Loading chunk data from a string
•	Placing new blocks
I have already started implementing the following features:
•	Checking for a block using a raycast
•	Destroying blocks
This is a very complicated and intriguing process, and I imagine I will be able to use the techniques from making this for many other purposes.