Update 28/08/2021 – 29/08/2021: Support for multiple block textures, explosions, attempts to optimise rendering
Added:
Set up chunk grid handler (whose name I have changed to LevelGrid since it’s shorter and snappier)
Support for blocks to have multiple texture types. My original plan was to have a system that procedurally stitches together different texture files when generating the new mesh, allowing me to have multiple separate texture files for ease of editing, but I decided this may be too time-consuming to implement. Currently, the textures are all included on one large file, and each block type has values to access particular parts of the texture file and obtain the necessary sections for UV mapping. I will need to eventually implement the more complex system, however, in case in future I need to add more complicated models for particular block types.
Explosions! I added in an Explosion class containing stats and a detonating function, allowing for explosions that deform terrain. I am quite pleased with this, although I might need to add some kind of randomising function to make the explosions more jagged looking, as the holes left behind the explosion are too uniform and spherical for my liking.

Worked on:
Tried to optimise mesh. Currently, the blocks at the ends of each chunk always render their exteriors, even if there’s a chunk adjacent to it. Since this mean lots of unnecessary polygons, I attempted to add code to the mesh generation so that if an adjacent block coordinate was outside the chunk, it would check if that coordinate’s world position was inside a valid chunk, and obtain the block data there instead. This has caused several bugs. When loading the level, I get null errors because the level generation code is refreshing each chunk at the same time it is being made, meaning the adjacent block generator is looking for chunks that haven’t been created yet. I tried adding a check to ensure the chunk was present before proceeding, but this caused the game to stall and become unresponsive when entering Play mode. The null check script is completely foolproof, so the problem may be with my adjacent chunk detection code. I realised that the sheer amount of block faces needing to be checked, combined with the code used for finding an adjacent chunk relying on several for loops, means I may be inadvertently multiplying the number of necessary calculations to a frankly ridiculous amount, causing a processing load too powerful for even my powerful enthusiast PC.
I also put a small amount of effort into sorting the files, as I am inviting other people to access and contribute to the project and I don’t want the workspace to look completely sloppy.

To implement next time:
Fix the problems mentioned when trying to fix the redundant mesh generation
Create more necessary block types
Figure out a reliable way to save and load chunk data.
Try creating useful level section prefabs that can be loaded, so the player can explore an actual environment rather than just digging into a solid cube.
Figure out exactly what kind of setup is needed for generating the level layouts, so the section prefabs can be arranged in an appropriate way.