Update 30/08/2021 pt. 2: Fixed the memory problem!
Turns out it was a problem that I’d thought I’d fixed already. When having code that checks adjacent blocks to render the appropriate faces, there is an if statement that ensures the adjacent block is not the same block as the previous one (as in, the exact identical block due to coordinates being clamped, not simply the same block type). I realised I had code that was checking the block structs themselves, but since different block structs can easily have the same data this was returning all kinds of false positives. I updated this check to instead ensure that the actual chunk coordinates are the same, and this seemed to fix the problem. I will need to add all the extra mesh optimisation code back in, that I originally wanted to include.
I also noticed that the explosion function isn’t very performant, and causes my reasonably powerful computer to freeze for about half a second.
One other thing I realised is that because the functions used to obtain block coordinates from world positions relies on simply converting the positions to local space, it will automatically account for chunks being rotated. I realised this means I should be able to freely rotate rooms without causing any trouble, and I don’t have to create any additional code for rotating room data.
I also need to fix the texture functionality, as something has broken it and now everything renders in solid black.