Update 31/08/2021 – Fixed textures, preparing for creation of actual prefabs
Restored the texture functionality to the destructible levels. I went through an embarrassing amount of troubleshooting before realising I’d accidentally commented out the one line of code that actually applied the UV data.

I attempted to re-implement the more optimised mesh generation, but after more troubles I decided to leave it for later.

Fixed a bug with the block destruction, where blocks that were invincible were still being destroyed.

Updated the Chunk class’ Refresh() function to only activate based on a bool, and altered the function for destroying blocks to act as a bool, which returns false if the block is empty or can’t be destroyed. I implemented this to ensure the chunk doesn’t refresh itself unnecessarily if no blocks were actually destroyed.

Added more functionality to facilitate easy editing of level sections. I altered the function for placing and destroying blocks to include code to scroll through all block types for the purpose of placement. I also took the actual controls and moved them into a different class. The code that detects blocks using the raycast now only exists for such a purpose, and I set it as such to make it more modular with any mechanic that involves interacting with chunks and blocks. For example, since we have decided the gameplay will be top-down and revolve primarily around smashing the floor with a hammer, there may not be a need for placing blocks in the final game. So I put it in a separate class for the purpose of modularity.

Made a secondary branch specifically for altering environment data.

I sorted many of the files. I moved the destructible environment code into a separate folder, as well as all of the art assets for the terrain.

Changed the product name.