Update 01/09/2021: Optimising chunk memory usage, working on a custom editor to reliably edit and save chunk layouts
Work start: 1pm
Work end: 9:30pm
Actual work start is roughly 2pm but I did a bit of work before that I wasn’t super good at getting the exact time frames for.

I found out that the block placement code doesn’t work if the player is placing the block on a face outside of a chunk, even if there is an adjacent chunk. This must be because there is no proper check for finding adjacent chunks. This is something I will need to work on.
Moved the Block struct into a separate script for ease of navigation, since it has a lot of functions now.
Converted the BlockInteraction class from a MonoBehaviour into a regular class, as there is no need for it to be attached to an object. It exists to be used by other functions.
Changed the Explosion class into a struct, since it is simply a collection of stats, with a built-in function for calculating the explosion, included within it so it can be ran without having to retype stats.
Updated the Chunk class so instead of generating a new mesh every time the chunk is refreshed, it declares a single Mesh class and updates the values in it. Since the existing class is being changed, no new data is actually being assigned to clutter up the memory. I also accomplished a similar feat with the chunk’s material – I noticed that when calculating UVs, I was referencing the chunk’s material instead of its shared material, which due to some quirk will create new material data every time it’s referenced. This is something I made sure to fix.

Put a sizeable amount of time and effort into making a system to build and save chunk data for room sections, which has not gone well. First, I attempted to simply directly alter the files in gameplay and drag them in as a prefab. This did not work, and the mesh was not retained. I then decided to make a custom editor named ChunkEditor so I could craft the levels in the editor rather than in gameplay. Although custom editors are not something I have much experience with, I found myself learning the systems rather quickly and putting together an interface. I currently have a function for easily selecting blocks to place from a dropdown, setting a size for generating explosions, and a function for resetting the chunk from scratch with a preset layout (useful if it’s currently empty, and I’m pleased with the failsafes I put in to ensure users don’t accidentally erase their work. I also intended for the user to use ScreenPointToRay() to place and destroy blocks by just clicking on the chunk in the editor’s scene view.

This, however, is still wracked with problems. The biggest one is that, due to a weird quirk in how custom editors save data, none of the block data is being saved. It appears that the only way to update an object properly in a custom editor is to directly alter certain values by creating, altering and applying serialised properties. This causes a lot of problems, as my destructible terrain system is primarily altered through functions which have many checks in order to ensure the terrain is updated in a responsible way. I can make changes while in the prefab editor, but they all revert back when I leave. I may be able to save the temporary block value and directly assign it when the editing is done, but this is something I will need to find out tomorrow.