Update 04/09/2021: Image-based saving and loading system, created but untested
Time: 12:30pm –3:30pm (minus about 1/12 hours of doing other things), 7:30pm – 8:30pm
Total time: 2h 30m
Performed work on the image-based save/load concept I described in the previous commit.
The saving function creates an array of 32-bit colour values, equivalent to the number of blocks in the chunk, plus 12. For each block in the chunk, the R, G, B and A values are used to store information. R represents the block’s ID, and G represents its remaining health. B and A will eventually be used for some kind of additional data, like directions or references to additional functionality. The remaining 12 colours in the array are for the dimensions of the array, the position, the rotation and the global scale of the transform. Since these four variables each consist of 3 floats (or ints in the case of the dimensions), and ints and floats have a size of four bytes (or 32 bits) each, this means that each value is big enough that only one can fit in a single colour value. As ints and floats consist of one 32-bit value, and a colour consists of four 8-bit values, I used the inbuilt C# BitConverter function to split up each int and float into four separate 8-bit values, and assign them as the R, G, B and A values of the colour. The code then obtains the square root of the colour array, and uses it as dimensions to generate a square texture with each colour in the array applied as one of the pixels.
To read the file, an in-built function is used to extract the pixels from the image and turn them back into an array of 32-bit colour values, which can then be cycled through to obtain the values therein. For the dimension and transform values, the R, G, B and A values of the 12 assigned colours are stitched back together into the original 32-bit numbers.
This functionality has not been tested yet (I’m not sure what the best GUI would be to allow the user to retrieve and input these image files), I am very confident in it being far better than my original save string idea.
I also created code that converts 3-dimensional array coordinates into single indexes, and vice versa. This will be helpful for creating functions that cycle through the grid without having 3 layers of for loops, as a single for loop that runs for a total of the multiplied dimensions of the array will run enough times to go through every value, and the index can be divided and separated to get the desired coordinates.