Update 07/09/2021: Implementing saving and loading function (although it doesn’t work properly yet), troubleshooting a mesh generation glitch
Times: 1:45pm – 2:45pm (mesh generation glitches), 2:45pm to 4:45pm (saving and loading)

I implemented a proper saving and loading function. It’s very buggy and does not work properly, but I am very sure I’m on the right track. When a player is editing a chunk in the game view, they can press F to save the chunk they are targeting, as an image file in the asset folder. When they press F over a chunk with the same name as one that is saved in the asset folder, the game will load the image from the file path and read the pixels to get the appropriate level data. There are a lot of bugs, with levels not coming out the exact way they were stored, but I feel I’m on the right track to making a usable editor for setting up levels.

I am also becoming incredibly aggravated by a recurring bug I have found, but have not figured out how to solve. Sometimes, when regenerating the chunk mesh, the resulting mesh will have a mismatch with there not being enough vertices for the amount of tri indexes. This results in the mesh not updating properly, with incomplete blocks and certain textures being applied to the wrong faces. One theory I have come up with is that this is some kind of memory issue – too much data is being created and destroyed by the garbage collector, so the game is somehow failing to save all of the correct information. I had previously encountered other glitches which I theorised could have existed for the same reason. While I am not completely convinced this is the problem, I also cannot think of any other reason it would be happening, as I am fully aware of every point in my code where the mesh data is being updated, and they are all working as intended.