Update 08/09/2021: Fixing mesh generation glitch, and attempting to optimise it
Time spent: 2:00pm – 5:30pm

I mentioned a problem earlier where the mesh generation would occasionally glitch out and display textures on many blocks in the wrong positions, and causing a very vague error about a mismatch of variables. After a lot of brainstorming (and asking on a forum for help), I learned that when updating certain mesh variables, it will immediately reconstruct the mesh. This can cause problems where if the new vertices or UVs are assigned before the new indices, the old indices will be used to generate a new mesh, often causing a mismatch. I tried assigning the indices first, which solved the problem but seemed to be worse for performance. I instead opted to simply implement the Mesh.Clear() function first to completely reset the mesh, which contrary to my expectations seemed to actually perform better than the first solution.

I also attempted to optimise the mesh further. The code relies on generating a ridiculous number of identical arrays which then get thrown out, and I attempted to remedy this problem by changing those identical arrays to static readonly ones, which could then be referenced without being created or destroyed. This worked for some of the arrays, but not for others, which produced glitches that I wasn’t sure how to solve. This is something I will need to fix next time.