Enemies are now destroyed when they are off the grid
Turns out I had to use a raycast to check if they were over a ground tile, because enemies usually don't use physics - they'd never actually fall down.
So if the rigidbody is currently kinematic (to allow players to knock enemies over holes), and they're not on top of the ground, they get destroyed.