Update 09/09/2021: Work on search function for saving and loading system.
Times: 3:30pm – 10pm

Today was not a great day for project progress. Despite spending a lot of time doing dev work on this assignment (as well as one of my personal projects, and putting a lot of effort into completing a practical test for a gamedev studio job application), my concentration was all over the place, and I didn’t achieve as much as I wanted to in the time I spent (although I think I did a pretty decent job with the job application test). I’m not sure if it was appropriate to spend the documented amount of time on the work I did, but it was something that needed to be created, and I like to imagine I learned a lesson on not to overthink mechanics.

I worked on a search function for the level editor, so I could efficiently save and load files. This function is mostly complete, and an important method to facilitate the generation of new levels, as the procedural nature of the chunk generation means level sections cannot be created in Unity’s regular editor. The current editor can edit data, and can technically save it (albeit with some bugs), but there is no practical function for saving data with specific file names, and loading from specific files (the way I selected names to test the saving and loading function was so terrible and inconvenient for regular use that I won’t even bother explaining it).

The search window is designed to be modular and work with many file types. I’ll describe how I intend for it to work (not accounting for elements that are currently not implemented or implemented incorrectly) The window is first populated by inputting an array of strings for the names of whatever values you want to pick. When the user types text into an input field, a for loop is used to go through all the possible entries, and add the indexes of the entries which contain the input field’s contents to a secondary list. This list will then be used to show the valid entries to select. When an entry from the searched list is selected, its index is returned and used to reference the original files being searched for. By demanding an input of strings, I can make the search window work with any variable with a name, and by making the output an index it is easy to reference the original array of items. Functions and secondary GUI elements can be adjusted accordingly, e.g. if the search function cannot find a valid entry, the buttons for loading and deleting files will be disabled because there is nothing selected to delete.

While I’m not pleased with the total amount of effort I made today, I feel I am headed in a good direction. I feel that just about every important element that I need to program is being implemented or fixed, and that once everything I have started is complete I will be done with the level generation system.