Update 14/09/2021 – 15/09/2021: Search function completed, level editor scrapped, replaced with slightly better but still buggy method, work on function for generating strings of rooms to make levels
Times on 14/09/2021: 1:15pm – 4:00pm
Completed the interface for saving and loading chunk data.
The chunk data is retrieved from the file properly, but the system is bugged and the data does not load in the correct way.
I need to replace this entire system because it’s buggy and overkill for what is actually needed and I wasted too much time because it turned out way more difficult than I could have predicted and due to a lot of outside factors I haven’t been performing at proper capacity. I need to fix this somehow.
Times on 15/09/2021: 11:30am – 6:30pm
Devised a similar but better system for generating rooms. Since the levels are technically only 2D and I need to get this done, I came up with a system that’s much easier to manually edit and therefore doesn’t require complicated editors. I created 2D image files representing top-down views of each room, where each colour represents a tile. Rather than trying to hijack the colour data values to store more data, this interface is technically less space efficient but can be easily edited without any kind of custom editor, since at a glance the pixels already present a rough approximation of the level layout. I used a new ScriptableObject type to reference particular colours, and specify what blocks and objects need to be spawned in that location.
I created several image files in an image editor, to represent different room and hallway layouts, and made a LevelGenerator script with an array that accepts them. This script will then generate a fixed number of randomly selected rooms, positioning and rotating them so the entries and exits line up to produce a cohesive route.
This system is showing more promise, but is very buggy, with rooms rotating in very strange ways and not positioning properly. While I’m worried I might be putting too much effort into this, I feel that it may be a good system. But I will need to fix the bugs and discuss it with the other team members.