Update 17/09/2021: Navmeshes generate properly, tile and object randomness in level generation, work on positioning rooms and musing about rotations
NavMesh work: 2:00pm – 4:00pm
Level generation: 4:00pm – 6:30pm

Properly updated the system for generating the NavMesh for the procedurally generating levels. I created a custom function named ‘AddSource’ that obtains a terrain object’s MeshFilter, and compiles a NavMeshBuildSource struct to add to the list. If a function for automatically generating bounds is enabled, the code will modify the existing bounds data to encompass the bounds of the new terrain mesh, by using GetComponent to find its MeshRenderer, and using its world accurate bounds.

I was worried using GetComponent that often would be poor for performance, but I was informed that it would not be a problem, and I realised that even if it was I was spending a bit too much time worrying about optimisation, when I just need to make it work. And fortunately, it does work. When generating the level, each time a room is created the code will use GetComponentsInChildren to find all the MeshFilters, and use AddSource on each one that is in the level geometry layer. With the NavMeshUpdateHandler completed, I removed the ChunkNavMeshHandler, since that was obsolete.

Updated import settings on the room files, since I believed that was the cause of some of the import problems. It didn’t fix everything, though.

Fixed the code for positioning new rooms, but not rotating them. The rooms will line up correctly as long as they generate in a straight line.

Added a variant of the LevelObjectFromPixel ScriptableObject, named VariedLevelObjectFromPixel. This variant lets the game randomly select one of multiple blocks and/or objects, which will be helpful for random variety in a roguelike, whether for visual aesthetic such as spawning different-looking blocks and walls, or gameplay variance such as rooms being filled with different enemy types.

I’m not sure how to handle rotation of objects, which will be necessary for certain assets such as walls and doors. I’m considering, when checking colours, only using the R, G and B values to determine the item type, and reserving the A value for determining rotation data, through means I haven’t quite figured out yet (maybe setting to one of the four cardinal directions based on the A value being within certain thresholds?) I will need to experiment with this.