Update 13/10/2021: Fixing room generation problems
Working on fixing individual room generation bugs: 11:00am – 2:00pm
Fixed a bug where certain tile types would give UnassignedReferenceException errors, messing up the generation of a room. This unfortunately took longer than expected, because I spent way too long on a complicated solution before realising the actual cause of the bug which could have been fixed by changing a single variable. This was incredibly aggravating, and now I’m stuck with a new system that has problems compared to the old one, and it would only be a waste of even more time to switch back to the old system but with the bugs fixed.
I also made a global static function for obtaining a random variable from an array, with code included for returning null if the size was less than zero. I am quite pleased with the clean nature of this function, as it will save me a lot of extra typing when obtaining variables from arrays.