Update 13/10/2021: File organisation and starting work on prefabs
Time spent: 4:30 – 5:30
Organised many of the team member’s files to be much easier to navigate to, sorted based on the asset types and optimised to have as few unnecessary folders and clicks as possible. It could probably be improved, but it’s presently much easier to navigate.
Created a proper pillar prefab from models given to me by the artists. I will do more work on this, but in my own branch.