Update 13/10/2021: Adding walls, a new room type, fixing some old rooms
Time spent: 5:30pm – 7:45pm
Tweaked the level generation algorithm and save files. If an object needs to have a specific rotation, that can be determined by the object’s alpha level. For each tile pixel, the alpha level is read as a 0-1 value and slotted into one of eight cardinal directions. This allows me to properly specify the directions of doors and walls.
Added proper wall prefabs to the level generator, which will randomly implement different wall variants for cosmetic effect.
Added a new room – the octagon room.
Fixed the square donut, small round and battle bridge rooms so they generate properly. I attempted to fix the straight connecting passage room, but that didn’t solve it.
Attempted to re-implement the functionality for rotating rooms for levels that were not straight arrangements of rooms, but this did not work.