Update 14/10/2021: Spawning in player, joining and rotating rooms properly
Work time: 11:00am – 1:30pm
Updated the ability for the player to spawn, into the level generation code.
Added the camera and HUD prefabs to the player prefab, to make the player work better as a single, modular package that can be spawned in.
Removed placeholder wall and door prefabs
Updated the connecting corner rooms so that the doors would spawn properly
Fixed code so now rooms generate in a consistent path and rotate properly. There is the issue of the rooms bending around into each other, which I am not sure how to fix yet, but this can be temporarily mitigated by removing connecting passages (since for some reason the straight passage keeps glitching).
Slightly updated the object spawning code in level generation so that the vertical position is adjusted properly to account for different transform positions. This code was present but did not work properly in certain circumstances, such as if the object did not directly have an attached mesh.
Added extra walls to the entry and exit door prefabs, so the player cannot slip out of the level while travelling between rooms.
Did some work on a RoomManager class to manage completion states, doors and enemies in each room. This is not fully implemented yet.