Update 14/10/2021 pt. 2: More general work on the stuff I was working on earlier
Made the doorways open by default so the player can explore the level without the need for finicky opening and closing code.
Added code so the player object is not a child of the room they spawn in.
Tried doing a bit more work on the RoomManager code but realised it wasn’t as important. It’s not necessary to the MVP for the player to be locked in rooms, just that they defeat all enemies in the level.