Update 15/10/2021: Merging systems together, trouble with scaling
Time: 11:00am – 6:00pm
Added the GameManager object to the GameLevel scene.
Added a place for the upgrade to appear at the end of the level, and implemented a LevelCompletionCheck class to ensure it only appears when all previous enemies are completed. The previous iteration of this code was worked into the obsolete enemy spawning script, and therefore had to be re-implemented on its own. The new system is not perfect but it works.
Every game process implemented so far occurs properly in conjunction with the procedurally generated levels. The player spawns in and can move through the level attacking enemies, the upgrade will spawn when the level is cleared of enemies, and the player will move to the next level upon collecting the upgrade. Nothing happens when the player finishes every level, but it seems according to the other team members’ code that nothing happened before, anyway. If the player dies, the game will reset appropriately.
The gameplay currently feels very awkward, but the reason is that in the flat test level, the blocky floor is scaled up by a factor of 2, so each tile is bigger and easier to aim at, and opens up more floor to drop enemies into. I did some work on automatically scaling up the level environment without affecting the player and enemies, but while I made some progress the system is not complete. I still have certain issues to find good solutions to.
Deleted the RoomManager class, since it wasn’t being used, had no valuable code in it and was just taking up space.